requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
supervision>=0.22.0
numpy>=1.24.0

paho-mqtt>=2.1.0

# Additional dependencies for the enhanced system
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
pyyaml>=6.0.0
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0
python-multipart>=0.0.6
jinja2>=3.1.0
